refactor: separate workflow queue and resource storage
Generated-By: pi 0.85.1 (https://pi.dev)
