+++ docs/2026-09-06-durable-execution-plan.md
@@ -152,4 +152,6 @@ All automated tests use temporary directories and deterministic model substitute
 - The server now derives response controls from the pending request kind. Waiting alone cannot expose `answer`. Terminal execution and durable pause keep their labels while worker or Pi activity is reported separately. Removed the remaining engine path that treated waiting as a terminal result. Typecheck, lint, 33 foundation tests, and nine live-view tests passed for this slice.
 - Response tools now resolve the exact durable request rather than the pending-view cache. The host checks live origin-session ownership before receipt lookup. Identical submissions adopt their saved result after settlement; conflicting payloads, stale coordinator epochs, other sessions, and the wrong response kind are rejected. Assistant response delivery has a separate private operation. Focused replay, authority, and active-time tests pass; typecheck and lint pass.
 - Every hosted execution launch now passes through one capacity check, including resource-manager reconciles, fresh starts, restart, resume, response validation, and expired interaction recovery. Pending launches count toward capacity. Start only reserves durable work. Removed direct response launches, the separate resource-manager capacity rule, and the in-memory pending-resume queue. Capacity is configurable; cancellation and renewal do not wait for it. Interactive reservations survive parking; headless work does not reserve a Pi session. Typecheck, lint, and all 37 focused queue, lifecycle, and capacity tests pass.
-- Queue-store extraction, alternate resource-runtime removal, timing, explicit model work, remaining client updates, and final validation remain in progress.
+- The complete deterministic suite passed all 1,186 tests after scheduler admission changed.
+- Moved run queue ownership into `src/workflows/queue.ts`. Resource-manager storage no longer exposes queue operations. Both domain stores reuse the existing SQLite connection and shared project/revision helpers. Removed the unused reserve-and-claim convenience API and old source-repair stubs. Queue, fencing, resource storage, extension, and scheduler tests passed except for one shared read-only connection case; the constructor now uses the connection's actual mode. All 20 tests in the follow-up view, protocol, effects, and hosted resource checks pass. Typecheck and lint pass.
+- Alternate resource-runtime removal, timing, explicit model work, remaining client updates, and final validation remain in progress.
