+++ test/queue-helpers.ts
@@ -1,0 +1,15 @@
+import type {
+  WorkflowRunQueueStore,
+  WorkflowRunReservationOptions,
+} from "../src/workflows/queue.js";
+
+/** A claimed queue fixture, without running the production scheduler. */
+export function claimTestRun(
+  queue: WorkflowRunQueueStore,
+  options: WorkflowRunReservationOptions & { claimToken: string; leaseMs: number },
+) {
+  queue.reserveWorkflowRun(options);
+  const claimed = queue.claimWorkflowRun(options);
+  if (claimed === undefined) throw new Error(`Test run could not be claimed: ${options.runId}`);
+  return claimed;
+}
